pyproject.toml
[tool.poetry]
name = "blog-backend"
version = "0.1.0"
description = ""
authors = ["endaaman <ken@endaaman.me>"]
readme = "README.md"
packages = [{include = "blog_backend"}]

[tool.taskipy.tasks]
dev = "uvicorn blog_backend:app --reload"
test = "pytest tests"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[tool.pytest.ini_options]
pythonpath = "project"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.20.0"
fastapi = "^0.92.0"
watchdog = "^2.2.1"
pylint = "^2.16.2"
pydantic = "^1.10.5"
python-dotenv = "^0.21.1"
taskipy = "^1.10.3"
aiofiles = "^23.1.0"
pyyaml = "^6.0"
cerberus = "^1.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
